fix(arnes): harden measurement collection boundaries
