+++ tooling/arnes/src/measure/repository.rs
@@ -22,13 +22,13 @@ pub fn observe(directory: &Path, root: String) -> RepositoryRecord {
 }
 
 pub fn protected_roots(directory: &Path, observed: Option<&Path>) -> Vec<PathBuf> {
-    let marker_root = directory
+    let mut roots = directory
         .ancestors()
         .filter(|ancestor| fs::symlink_metadata(ancestor.join(".git")).is_ok())
         .last()
-        .unwrap_or(directory)
-        .to_owned();
-    let mut roots = vec![marker_root];
+        .map(Path::to_owned)
+        .into_iter()
+        .collect::<Vec<_>>();
     if let Some(observed) = observed.filter(|observed| !roots.iter().any(|root| root == observed)) {
         roots.push(observed.to_owned());
     }
